+++ src/main/java/com/shop/farmers/boundedContext/order/entity/Order.java
@@ -53,8 +53,8 @@ public class Order extends BaseEntity {
         return order;
     }
 
-    public int getTotalPrice(){
-        int totalPrice = 0;
+    public Long getTotalPrice(){
+        long totalPrice = 0;
         for(OrderItem orderItem : orderItems){
             totalPrice += orderItem.getTotalPrice();
         }
